use already build envd for integrtations tests template build
currently there is copy paste of already build envd from our public
bucket, but in case we want to test changes in envd itself we should
build it ourself in every integrations tests run.

for back compatibility we are still copying legacy version that must be
present in sandbox. This is taken from archive.