Update Makefile - properly handle error codes in plan part
The terraform exits correctly with error code 2, which means there are actually changes. That is ok, no need to for Makefile to end with error code as well. This is what we want (I believe). 

To perform exactly these actions, run the following command to apply:
    terraform apply ".tfplan.prod"
make: *** [Makefile:84: plan] Error 2

vs

To perform exactly these actions, run the following command to apply:
    terraform apply ".tfplan.prod"
Changes detected.