Nomad nodes discovery is now using shared logic for job allocations
Client proxy previously used allocations job search but only for one
jobs with jobs prefix name. This was not suitable for running in
environment where it should be used for discover of both orchestrator
and template manager instances.

New filter logic search for template managers on build nodes and
orchestrators on client default nodes. Now ENV variable for job prefix
is no longer needed as its baked in discovery logic itself.

Logic was also moved from client proxy package to shared as it will be
later used for same job in API, that now uses edge api for local cluster
template builders disocvery. This will be changed in following changes.