Rename structs handling instance info
Lot of structs and variables is prefixed orchestrator even its already
in package name so its not needed.

Rename variables on structs/struct itself to start with lower letter to
not export them to other packages if not really needed.

Instance gRPC client is also made internal and we now have two function
for exporting raw gRPC connection used in gRPC proxy and gRPC info
service client that is used in few places.