Edge pool renaming and cleanup
Moved instance client holding gRPC connection and gRPC info client into
separated file to make it cleaner and easier to work with.

Orchestrator filename is renamed to instance and
`NewOrchestratorInstance` is renamed to `newInstance` as orchestrator is
duplicated here (package name) and dont need to be used ourside of
package.