Introduced cluster resources provider
This is interface ment for unify all calls to resources we want to query
from cluster (sandbox logs, metrics, build logs etc).

This commit introduces basic implementation and getter on cluster that
will be later used on all places where we need to query sandbox/build
logs and/or metrics.

Later this same interface will be used for local cluster to query local
Loki and ClickHouse to provide same data. This way we should have
unified interface.